added drawer component
