+++ src/components/table/MembersTable.jsx
@@ -1,25 +1,40 @@
+import { TableBody, TableRow } from "@mui/material";
 import {
-  Avatar,
-  styled,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Tooltip,
-  TableFooter,
-  TablePagination,
-  tooltipClasses,
-  Switch,
-  Box,
-} from "@mui/material";
+  ReportingList,
+  TableDiv,
+  StyledTableContainer,
+  StyledBottomTableContainer,
+  StyledTablePagination,
+  StyledTableFooter,
+  StyledTableRow,
+  MemberProfile,
+  MoreTag,
+  SignalsAvatar,
+  StyledTableHead,
+  StyledTableHeading,
+  StyledTableCell,
+  ToolTipContent,
+  SignalList,
+  LightTooltip,
+  StickyHeading,
+  ProfileAvatar,
+  StickyCell,
+  PencilIcon,
+  ActionContainer,
+  ActionStickyContainerSeparator,
+  ActionPaginationContainerSeparator,
+  ActionHeaderContainerSeparator,
+  ActionContent,
+  StyledTable,
+} from "./Table.styles";
+
 import ReactSpeedometer from "react-d3-speedometer";
-import CreateOutlinedIcon from "@mui/icons-material/CreateOutlined";
 import Table from "@mui/material/Table";
 import IOSSwitch from "../switch/Switch";
-import PencilIconSvg from "../../assets/icons/pencil.svg";
 import StyledButton from "../button/Button";
 import { useEffect, useState } from "react";
+import Drawer from "../drawer/Drawer";
+import StyledDrawer from "../drawer/Drawer";
 const MembersTable = ({
   headings,
   stickyheadings,
@@ -30,6 +45,7 @@ const MembersTable = ({
 }) => {
   const [filteredRows, setFilteredRows] = useState([]);
   const [currPage, setCurrPage] = useState(0);
+  const [isEditMemberDrawerOpen, setIsEditMemberDrawerOpen] = useState(true);
   const [rowsPerPage, setRowsPerPage] = useState(10);
   const [modifiedRows, setModifiedRows] = useState([]);
   useEffect(() => {
@@ -210,7 +226,11 @@ const MembersTable = ({
                         >
                           Add Feedback
                         </StyledButton>
-                        <PencilIcon />
+                        <PencilIcon
+                          onClick={() => {
+                            setIsEditMemberDrawerOpen(true);
+                          }}
+                        />
                       </ActionContent>
                     </ActionContainer>
                   </StickyCell>
@@ -240,221 +260,16 @@ const MembersTable = ({
           </StyledTableFooter>
         </Table>
       </StyledBottomTableContainer>
+      <StyledDrawer
+        content={"Hi"}
+        anchor={"right"}
+        onClose={() => {
+          setIsEditMemberDrawerOpen(false);
+        }}
+        open={isEditMemberDrawerOpen}
+      />
     </TableDiv>
   );
 };
-
-const ReportingList = styled("div")({
-  display: "flex",
-  flexDirection: "column",
-  gap: "10px",
-  padding: "5px",
-});
-
-const TableDiv = styled(Box)({
-  position: "relative",
-  borderRadius: "10px",
-  backgroundColor: "white",
-  borderLeft: "solid white 15px",
-  boxShadow: "0px 0px 10px rgba(0, 0, 0, 0.12) ",
-});
-
-const StyledTableContainer = styled(TableContainer)({
-  backgroundColor: "white",
-  borderRadius: "10px  10px 0px 0px ",
-  position: "relative",
-  fontFamily: "Poppins",
-  minHeight: "40rem",
-  maxHeight: "40rem",
-  scrollSnapType: "both mandatory",
-  scrollSnapAlign: "start",
-});
-
-const StyledBottomTableContainer = styled(TableContainer)({
-  backgroundColor: "white",
-
-  borderRadius: "0px 0px 10px  10px",
-  fontFamily: "Poppins",
-  zIndex: 10,
-  position: "relative",
-});
-
-const StyledTablePagination = styled(TablePagination)({
-  "& .MuiTablePagination-root": {
-    fontFamily: "Poppins",
-  },
-  "& .MuiTablePagination-selectLabel": {
-    fontFamily: "Poppins",
-  },
-  "& .MuiTablePagination-select": {
-    fontFamily: "Poppins",
-  },
-  "& .MuiTablePagination-displayedRows": {
-    fontFamily: "Poppins",
-  },
-});
-
-const StyledTableFooter = styled(TableFooter)({});
-
-const StyledTableRow = styled(TableRow)({});
-
-const MemberProfile = styled("div")(({ fontSize }) => ({
-  display: "flex",
-  alignItems: "center",
-  gap: "5px",
-  fontSize,
-  width: "100%",
-}));
-
-const MoreTag = styled("div")({
-  color: "#49C792",
-});
-
-const SignalsAvatar = styled(Avatar)(({ bgColor, color }) => ({
-  backgroundColor: bgColor,
-  color: color,
-  padding: 0,
-  width: "30px",
-  height: "30px",
-  fontSize: "small",
-}));
-
-const StyledTableHead = styled(TableHead)({
-  fontWeight: "bold",
-  fontFamily: "Poppins",
-});
-
-const StyledTableHeading = styled(TableCell)(({ minWidth }) => ({
-  fontWeight: "500",
-  fontSize: "1.07em",
-  fontFamily: "Poppins",
-  minWidth: minWidth || 130,
-  position: "sticky",
-  top: 0,
-  backgroundColor: "white",
-  zIndex: 1,
-  borderBottom: "none",
-}));
-
-const StyledTableCell = styled(TableCell)(({ minWidth }) => ({
-  fontSize: "1.07em",
-  borderBottom: "none",
-  padding: "0px 15px",
-  cursor: "pointer",
-  fontFamily: "Poppins",
-  minWidth: minWidth || "9rem",
-}));
-
-const ToolTipContent = styled("div")({
-  display: "flex",
-  flexDirection: "column",
-});
-
-const SignalList = styled("div")({
-  display: "flex",
-  alignItems: "center",
-  gap: "10px",
-});
-
-const LightTooltip = styled(({ className, fontSize, ...props }) => (
-  <Tooltip {...props} classes={{ popper: className }} />
-))(({ theme, fontSize }) => ({
-  [`& .${tooltipClasses.tooltip}`]: {
-    backgroundColor: theme.palette.common.white,
-    color: "rgba(0, 0, 0, 0.87)",
-    boxShadow: "0px 0px 10px rgba(0, 0, 0, 0.12) ",
-    fontSize: fontSize || 11,
-  },
-  [`& .${tooltipClasses.arrow}`]: {
-    color: theme.palette.common.white,
-  },
-}));
-
-const StickyHeading = styled(TableCell)({
-  position: "sticky",
-  right: 0,
-  zIndex: 2,
-  fontWeight: "500",
-  fontSize: "18px",
-  fontFamily: "Poppins",
-  borderBottom: "none",
-  minWidth: 150,
-  backgroundColor: "white",
-  position: "sticky",
-  bottom: 0,
-  top: 0,
-});
-
-const ProfileAvatar = styled(Avatar)(({ height, width, fontSize }) => ({
-  height,
-  width,
-  fontSize,
-}));
-
-const StickyCell = styled(TableCell)({
-  position: "sticky",
-  right: 0,
-  fontSize: "16px",
-  zIndex: 1,
-  minWidth: 200,
-  cursor: "pointer",
-  backgroundColor: "white",
-  borderBottom: "none",
-  padding: "20px 10px",
-  fontFamily: "Poppins",
-});
-
-const PencilIcon = styled((props) => <img src={PencilIconSvg} {...props} />)({
-  color: "#49C792",
-  backgroundColor: "#EEFBF6",
-  padding: "2px",
-  borderRadius: "100px",
-});
-
-const ActionContainer = styled("div")({
-  display: "flex",
-  alignItems: "center",
-  gap: "10px",
-  position: "relative",
-  width: "100%",
-});
-
-const ActionStickyContainerSeparator = styled("div")({
-  borderLeft: "solid rgba(0, 0, 0, 0.12) 1px",
-  height: "100%",
-  position: "absolute",
-  zIndex: 4,
-  right: 0,
-  transform: "translate(-220px,0%)",
-});
-
-const ActionPaginationContainerSeparator = styled("div")({
-  width: "calc(100% + 15px)",
-  borderBottom: "solid rgba(0, 0, 0, 0.12) 1px",
-  position: "absolute",
-  bottom: 0,
-  transform: "translate(-15px,-52px)",
-  zIndex: 11,
-});
-
-const ActionHeaderContainerSeparator = styled("div")({
-  width: "calc(100% + 15px)",
-  borderBottom: "solid rgba(0, 0, 0, 0.12) 1px",
-  position: "absolute",
-  transform: "translate(-15px,55px)",
-  zIndex: 10,
-});
-
-const ActionContent = styled("div")({
-  display: "flex",
-  alignItems: "center",
-  gap: "10px",
-  justifyContent: "center",
-  flex: 1,
-});
-
-const StyledTable = styled(Table)({
-  position: "relative",
-});
 
 export default MembersTable;
